# go/model_todo.go
/*
 * Todo app OAS
 *
 * OpenApi specification for a todo application
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type Todo struct {

	Id int32 `json:"id,omitempty"`

	Description string `json:"description"`

	Title string `json:"title"`

	Reporter *User `json:"reporter"`

	Assignee *User `json:"assignee,omitempty"`

	Responsibility string `json:"responsibility"`

	Status string `json:"status"`

	Priority string `json:"priority"`

	CreatedAt time.Time `json:"created_at,omitempty"`

	CompletedAt time.Time `json:"completed_at,omitempty"`
}