# go/model_user.go
/*
 * Todo app OAS
 *
 * OpenApi specification for a todo application
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type User struct {

	Id int32 `json:"id,omitempty"`

	Firstname string `json:"firstname"`

	Surname string `json:"surname"`

	Email string `json:"email"`

	Role string `json:"role"`
}